Improve Google role setup step
