+++ app/workflow/step-ids.ts
@@ -6,8 +6,7 @@ export const StepId = {
   VerifyPrimaryDomain: "verify-primary-domain",
   CreateAutomationOU: "create-automation-ou",
   CreateServiceUser: "create-service-user",
-  CreateCustomAdminRole: "create-custom-admin-role",
-  AssignRoleToUser: "assign-role-to-user",
+  CreateAdminRoleAndAssignUser: "create-admin-role-and-assign-user",
   ConfigureGoogleSamlProfile: "configure-google-saml-profile",
   CreateMicrosoftApps: "create-microsoft-apps",
   ConfigureMicrosoftSyncAndSso: "configure-microsoft-sync-and-sso",
