+++ app/workflow/steps/AGENTS.md
@@ -161,6 +161,7 @@ Response `200 OK` (OU exists)
 
 - `googleAccessToken`
 - `isDomainVerified`
+- `provisioningUserId`
 
 #### Request
 
@@ -251,11 +252,11 @@ Content-Type: application/json
 // No vars provided; step ensures OU existence only
 ```
 
-## Step 4: `createCustomAdminRole`
+## Step 4: `createAdminRoleAndAssignUser`
 
 ### Purpose
 
-Ensure custom admin role `Microsoft Entra Provisioning` exists with correct privileges.
+Ensure custom admin role `Microsoft Entra Provisioning` exists with correct privileges and is assigned to the provisioning user.
 
 ### State Check
 
@@ -270,12 +271,18 @@ Authorization: Bearer {googleAccessToken}
 
 ```json
 {
+  "kind": "admin#directory#roles",
   "items": [
     {
-      "roleId": "R123",
+      "roleId": "91447453409035723",
       "roleName": "Microsoft Entra Provisioning",
-      "rolePrivileges": [ { "serviceId": "svc", ... }, ... ]
-    }, ...
+      "roleDescription": "Custom role for Microsoft provisioning",
+      "rolePrivileges": [
+        { "privilegeName": "USERS_CREATE", "serviceId": "00haapch16h1ysv" },
+        { "privilegeName": "USERS_RETRIEVE", "serviceId": "00haapch16h1ysv" },
+        { "privilegeName": "USERS_UPDATE", "serviceId": "00haapch16h1ysv" }
+      ]
+    }
   ]
 }
 ```
@@ -322,58 +329,22 @@ Content-Type: application/json
   "roleName": "Microsoft Entra Provisioning",
   "roleDescription": "Custom role for Microsoft provisioning",
   "rolePrivileges": [
+    { "serviceId": "{directoryServiceId}", "privilegeName": "ORGANIZATION_UNITS_RETRIEVE" },
     { "serviceId": "{directoryServiceId}", "privilegeName": "USERS_RETRIEVE" },
     { "serviceId": "{directoryServiceId}", "privilegeName": "USERS_CREATE" },
-    { "serviceId": "{directoryServiceId}", "privilegeName": "USERS_UPDATE" }
+    { "serviceId": "{directoryServiceId}", "privilegeName": "USERS_UPDATE" },
+    { "serviceId": "{directoryServiceId}", "privilegeName": "GROUPS_ALL" }
   ]
 }
 ```
 
 #### Expected Responses
 
-- `201 Created`: Role created
+- `200 OK`: Role created
 - `409 Conflict`: Role already exists (acceptable)
 - `400/403`: error
 
-#### Variables Extracted on Success
-
-```ts
-adminRoleId = .roleId
-```
-
-## Step 5: `assignRoleToUser`
-
-### Purpose
-
-Ensure the custom role is assigned to the service user.
-
-### State Check
-
-#### Request
-
-```http
-GET https://admin.googleapis.com/admin/directory/v1/customer/my_customer/roleassignments?roleId={adminRoleId}&userKey={provisioningUserId}
-Authorization: Bearer {googleAccessToken}
-```
-
-#### Success Response (`200 OK`)
-
-```json
-{ "items": [ { ... } ] }
-```
-
-#### Completion Criteria
-
-`items` array length >= 1
-
-### Execution
-
-#### Prerequisites
-
-- `googleAccessToken`
-- `adminRoleId`, `provisioningUserId`, `isDomainVerified`
-
-#### Request
+#### Request 3: Assign Role
 
 ```http
 POST https://admin.googleapis.com/admin/directory/v1/customer/my_customer/roleassignments
@@ -389,10 +360,37 @@ Content-Type: application/json
 
 #### Expected Responses
 
-- `201 Created` or `409 Conflict`
+- `200 OK` or `409 Conflict`
 - `400/404/403`: error
 
-## Step 6: `configureGoogleSamlProfile`
+Example success:
+
+```json
+{
+  "roleAssignmentId": "91447453409034880",
+  "roleId": "91447453409035734",
+  "assignedTo": "103898700330622175095"
+}
+```
+
+Conflict example:
+
+```json
+{
+  "error": {
+    "code": 409,
+    "message": "Role assignment already exists for the role"
+  }
+}
+```
+
+#### Variables Extracted on Success
+
+```ts
+adminRoleId = .roleId
+```
+
+## Step 5: `configureGoogleSamlProfile`
 
 ### Purpose
